+++ hutool-crypto/src/main/java/cn/hutool/crypto/digest/mac/MacEngineFactory.java
@@ -1,9 +1,9 @@
 package cn.hutool.crypto.digest.mac;
 
-import javax.crypto.SecretKey;
-
 import cn.hutool.crypto.SmUtil;
 import cn.hutool.crypto.digest.HmacAlgorithm;
+
+import java.security.Key;
 
 /**
  * {@link MacEngine} 实现工厂类
@@ -19,7 +19,7 @@ public class MacEngineFactory {
 	 * @param key 密钥
 	 * @return {@link MacEngine}
 	 */
-	public static MacEngine createEngine(String algorithm, SecretKey key) {
+	public static MacEngine createEngine(String algorithm, Key key) {
 		if(algorithm.equalsIgnoreCase(HmacAlgorithm.HmacSM3.getValue())) {
 			// HmacSM3算法是BC库实现的
 			return SmUtil.createHmacSm3Engine(key.getEncoded());